Eval/torchllms/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "torchllms"
version = "0.0.1"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
]
dependencies = [
    "blessed",
    "datasets",
    "immutabledict",
    "jsonlines",
    "langdetect",
    "loguru",
    "plotext",
    "protobuf",
    "pydantic",
    "pygments",
    "python-dotenv",
    "pyyaml",
    "rich",
    "sentencepiece",
    "torch",
    "transformers",
    "tqdm",
]

[tool.setuptools.packages.find]
include = ["torchllms*"]

[tool.wheel]
include = ["torchllms*"]